Dev full restfull api comment

+++ server/src/controllers/commentController.js
@@ -8,6 +8,7 @@ const commentModel = require('../models/commentModel');
 
 //Import valid
 const commentValid = require('../utils/commentValid');
+const { default: Item } = require('antd/lib/list/Item');
 
 const commentController = {
   create: catchAsyncError(async (req, res) => {
@@ -20,17 +21,6 @@ const commentController = {
         statusCode: 400,
         err: msgError,
       });
-
-    //Check exist id comment parent
-    if (comment_id) {
-      //Check exist parent comment
-      const parentComment = await commentModel.findById(comment_id);
-      if (!parentComment)
-        return res.status(400).json({
-          statusCode: 400,
-          err: 'Parent comment does not exist in system.',
-        });
-    }
 
     //Check exist user
     const user = await userModel.findById(user_id);
@@ -48,13 +38,34 @@ const commentController = {
         err: 'Idea does not exist in system.',
       });
 
-    //Save comment
-    await commentModel.create({
-      content,
-      user_id,
-      idea_id,
-      comment_id: comment_id ? comment_id : null,
-    });
+    //Check exist id comment parent
+    if (comment_id) {
+      //Check exist parent comment
+      const parentComment = await commentModel.findById(comment_id);
+      if (!parentComment)
+        return res.status(400).json({
+          statusCode: 400,
+          err: 'Parent comment does not exist in system.',
+        });
+
+      //Update replies when have parent comment
+      parentComment.replies = [
+        {
+          user_id,
+          content,
+        },
+        ...parentComment.replies,
+      ];
+
+      await parentComment.save();
+    } else {
+      //Save comment
+      await commentModel.create({
+        content,
+        user_id,
+        idea_id,
+      });
+    }
 
     res.status(200).json({
       statusCode: 200,
@@ -65,7 +76,7 @@ const commentController = {
   update: catchAsyncError(async (req, res) => {
     const { comment_id } = req.params;
 
-    const { content } = req.body;
+    const { content, reply_id } = req.body;
 
     //Check valid data
     const msgError = commentValid.updateCommentValid({ content });
@@ -75,17 +86,38 @@ const commentController = {
         err: msgError,
       });
 
-    //Check exist comment
-    const comment = await commentModel.findById(comment_id);
-    if (!comment)
-      return res.status(400).json({
-        statusCode: 400,
-        err: 'Comment does not exist in system.',
-      });
+    //Check exist id comment parent
+    if (reply_id) {
+      //Check exist parent comment
+      const comment = await commentModel.findById(comment_id);
+      if (!comment)
+        return res.status(400).json({
+          statusCode: 400,
+          err: 'Comment does not exist in system.',
+        });
 
-    //Update comment
-    comment.content = content;
-    await comment.save();
+      //Update replies when have parent comment
+      await commentModel.update(
+        { _id: comment_id, 'replies._id': reply_id },
+        {
+          $set: {
+            'replies.$.content': content,
+          },
+        }
+      );
+    } else {
+      //Check exist comment
+      const comment = await commentModel.findById(comment_id);
+      if (!comment)
+        return res.status(400).json({
+          statusCode: 400,
+          err: 'Comment does not exist in system.',
+        });
+
+      //Update comment
+      comment.content = content;
+      await comment.save();
+    }
 
     return res.status(200).json({
       statusCode: 200,
@@ -95,17 +127,35 @@ const commentController = {
 
   delete: catchAsyncError(async (req, res) => {
     const { comment_id } = req.params;
+    const { reply_id } = req.query;
 
-    //Check exist comment
-    const comment = await commentModel.findById(comment_id);
-    if (!comment)
-      return res.status(400).json({
-        statusCode: 400,
-        err: 'Comment does not exist in system.',
+    if (reply_id) {
+      //Check exist parent comment
+      const comment = await commentModel.findById(comment_id);
+      if (!comment)
+        return res.status(400).json({
+          statusCode: 400,
+          err: 'Comment does not exist in system.',
+        });
+
+      //Delete replies when have parent comment
+      await commentModel.findByIdAndUpdate(comment_id, {
+        $pull: {
+          replies: { _id: reply_id },
+        },
       });
+    } else {
+      //Check exist comment
+      const comment = await commentModel.findById(comment_id);
+      if (!comment)
+        return res.status(400).json({
+          statusCode: 400,
+          err: 'Comment does not exist in system.',
+        });
 
-    //Delete comment
-    commentModel.findByIdAndDelete(comment_id)
+      //Delete comment
+      await commentModel.findByIdAndDelete(comment_id);
+    }
 
     return res.status(200).json({
       statusCode: 200,
